move Concatenate to util/concatenate
